fix: moved context to independent file
