+++ examples/tsconfig.json
@@ -23,6 +23,6 @@
     ],
     "exclude": [
         "./dist/",
-        "../src"
+        "../src/"
     ]
 }
