+++ src/cardboard.ts
@@ -1,3 +1,4 @@
+import * as _context from './context.js';
 import * as _tag from './tag.js';
 import * as _state from './state.js';
 import * as _css from './css-generator.js';
@@ -6,8 +7,8 @@ import * as _text from './text.js';
 import * as _events from './events.js';
 import * as _lifecycle from './lifecycle.js';
 import * as _observables from './observables.js';
-import * as _routing from './ext/routing.js';
 
+export * from './context.js';
 export * from './tag.js';
 export * from './state.js';
 export * from './css-generator.js';
@@ -17,7 +18,6 @@ export * from './events.js';
 export * from './each.js';
 export * from './lifecycle.js';
 export * from './observables.js';
-export * from './ext/routing.js';
 export type * from './types';
 
 // Import the version from package.json
@@ -29,8 +29,8 @@ export const Cardboard = {
   ..._util,
   ..._text,
   ..._events,
-  ..._routing,
   ..._observables,
   ..._lifecycle,
+  ..._context,
   version: '0.0.4',
 };
